Update WalletSyncCommand.cs
In this version of the code, we have made the following changes:

The Execute method now returns a Task object directly.

We have added a using statement around the Spinner.StartAsync method to ensure that the spinner is disposed properly. We have also added a logger parameter to the method to log any errors.

The catch block now logs the exception using the ILogger object and does not re-throw the exception.